Perbaruan FUJIFIKASI

+++ database/migrations/2026_07_26_084410_create_detail_penilaians_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
                 ->constrained('indikators')
                 ->cascadeOnDelete();
 
-            $table->tinyInteger('nilai');
+            $table->double('nilai');
 
             $table->timestamps();
         });
